+++ src/components/mypage/admin/ReviewListModal/ReviewListModal.module.css
@@ -1,0 +1,47 @@
+/* review modal */
+.review-count {
+  display: block;
+  width: fit-content;
+  margin: 0 1.5rem 1rem auto;
+  font-size: 0.875rem;
+}
+
+.review {
+  width: 95%;
+  margin: 0 auto;
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  gap: 0.875rem;
+  margin-bottom: 1.25rem;
+}
+
+.review-contents {
+  padding: 1.5rem 1.875rem 2rem 1.875rem;
+  border: 1px solid var(--gray100-color);
+  border-radius: 10px;
+  position: relative;
+  flex-grow: 1;
+}
+
+.review__nickname {
+  display: block;
+  margin-bottom: 0.625rem;
+  font-size: 0.875rem;
+  font-weight: bold;
+}
+
+.review__content {
+  font-size: 0.75rem;
+  color: var(--gray400-color);
+  word-break: break-all;
+}
+
+.review__date {
+  position: absolute;
+  bottom: 0.8rem;
+  right: 0.9375rem;
+  font-size: 0.625rem;
+  font-weight: bold;
+  color: var(--gray300-color);
+}
