+++ src/components/mypage/admin/ShowList/ShowList.module.css
@@ -1,0 +1,30 @@
+.list {
+  box-sizing: border-box;
+  width: 100%;
+  height: 60vh;
+  margin-bottom: 1.25rem;
+  border: 1px solid var(--gray300-color);
+  border-radius: 5px;
+  background-color: var(--gray100-color);
+  overflow-y: auto;
+}
+
+.list-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 1rem;
+  padding: 1.125rem 1.5rem;
+  border: 1px solid var(--gray100-color);
+  background-color: white;
+  text-align: center;
+  word-break: keep-all;
+  border-radius: 5px;
+  font-size: 0.75rem;
+}
+
+@media (max-width: 520px) {
+  .list-row {
+    flex-direction: column;
+  }
+}
