+++ src/components/mypage/index.tsx
@@ -8,6 +8,9 @@ import ReservationItem from "./user/ReservationItem/ReservationItem";
 import ReservationUserModal from "./user/ReservationUserModal/ReservationUserModal";
 import ReviewItem from "./user/ReviewItem/ReviewItem";
 import StoryItem from "./user/StoryItem/StoryItem";
+import ShowList from "./admin/ShowList/ShowList";
+import ShowListItem from "./admin/ShowListItem/ShowListItem";
+import ReviewListModal from "./admin/ReviewListModal/ReviewListModal";
 
 export {
   ReservationListModal,
@@ -20,4 +23,7 @@ export {
   ShowInfoInputsSection,
   ShowResInfoInputsSection,
   ImageInputSection,
+  ShowList,
+  ShowListItem,
+  ReviewListModal,
 };
