--- src/pages/Mypage/admin/PostManage/PostManagePage.module.css
@@ -1,130 +1,0 @@
-.list {
-  box-sizing: border-box;
-  width: 100%;
-  height: 60vh;
-  margin-bottom: 1.25rem;
-  border: 1px solid var(--gray300-color);
-  border-radius: 5px;
-  background-color: var(--gray100-color);
-  overflow-y: auto;
-}
-
-.list-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 1rem;
-  padding: 1.125rem 1.5rem;
-  border: 1px solid var(--gray100-color);
-  background-color: white;
-  text-align: center;
-  word-break: keep-all;
-  border-radius: 5px;
-  font-size: 0.75rem;
-}
-
-.list-row-left {
-  text-align: start;
-}
-
-.list-row-right {
-  display: flex;
-  flex-shrink: 0;
-  gap: 0.8rem;
-  align-items: center;
-}
-
-@media (max-width: 520px) {
-  .list-row {
-    flex-direction: column;
-  }
-  .list-row-left {
-    text-align: center;
-  }
-  .list-row-right {
-    flex-basis: auto;
-  }
-}
-
-.list-item__title {
-  margin-bottom: 0.5rem;
-  font-size: 0.875rem;
-}
-
-.list-item__date {
-  color: var(--gray300-color);
-}
-
-.list-row-right button {
-  box-sizing: border-box;
-  font-size: 0.75rem;
-  width: fit-content;
-  word-break: keep-all;
-  padding: 0.8rem;
-  margin: auto;
-}
-
-svg.cmt-icon {
-  margin-right: 0.4rem;
-  height: 0.8rem;
-}
-
-.like-cnt {
-  display: flex;
-  align-items: center;
-}
-.like-cnt svg {
-  height: 0.8rem;
-  margin-right: 0.2rem;
-}
-.like-cnt svg path {
-  stroke: var(--gray400-color);
-}
-
-/* review modal */
-.review-count {
-  display: block;
-  width: fit-content;
-  margin: 0 1.5rem 1rem auto;
-  font-size: 0.875rem;
-}
-
-.review {
-  width: 95%;
-  margin: 0 auto;
-  box-sizing: border-box;
-  display: flex;
-  align-items: center;
-  gap: 0.875rem;
-  margin-bottom: 1.25rem;
-}
-
-.review-contents {
-  padding: 1.5rem 1.875rem 2rem 1.875rem;
-  border: 1px solid var(--gray100-color);
-  border-radius: 10px;
-  position: relative;
-  flex-grow: 1;
-}
-
-.review__nickname {
-  display: block;
-  margin-bottom: 0.625rem;
-  font-size: 0.875rem;
-  font-weight: bold;
-}
-
-.review__content {
-  font-size: 0.75rem;
-  color: var(--gray400-color);
-  word-break: break-all;
-}
-
-.review__date {
-  position: absolute;
-  bottom: 0.8rem;
-  right: 0.9375rem;
-  font-size: 0.625rem;
-  font-weight: bold;
-  color: var(--gray300-color);
-}
